Fix name of method for returning connection class
This method was previously called `connection_classes` but it can only
return one class. This is a private API so it doesn't matter for apps,
but I need to use it for a followup PR and was confused that it implied
it could return multiple values but only ever returned one.