Fix flakey test in Active Record dirty_test.rb
Example failure: https://buildkite.com/rails/rails/builds/82957#44f56e67-28e9-4193-9175-695e300a9756/1110-1120

The test asserts that `manufactured_at` defaults to `CURRENT_TIMESTAMP`
from the database by comparing it to `Time.now`, with both timestamps
truncated to seconds precision.  However, if the record is saved during
the final milliseconds of a second, and `Time.now` is evaluated during
the first milliseconds of the next second, the truncated timestamps will
still differ.

This commit changes the assertion to instead check that the timestamps
are within 1 second of each other.