Remove memoization for limited_count
since if a relation's records are not loaded and @limited_count gets set, when some of the records are subsequently deleted then the memoized value for @limited_count is inconsistent with the value returned by a call to size. In such a case calls to one? and many? can be incorrect.