Don't match "time" inside SET()/ENUM()
We previously relied upon definition order to prevent this, but now
we're defining these types in a child TypeMap, so that won't work.

The loose patterns we define here seem like they'd be in general danger
of this sort of issue, but for now we can just address the immediately
affected ones.