Allow localhost with a port by default in development
Only "localhost" and ".localhost" were allowed by default in
development, but the port is also part of the header that is sent.  It
is a bad idea to try parsing whatever stuff is in the HTTP header, so
this commit specifically adds localhost with a port in the "allowed
hosts" list in development.