Fix scaffold index.html.erb variable name
Due to a conflict between #43846 and #43611, the variable name generated
inside the `each` block did not match the block parameter name for
namespaced models.  For example, if the block parameter name was
`admin_role`, the variable name inside the block would be just `role`.

This commit changes `index.html.erb` to use the correct variable name,
and adds much more test coverage.

This commit also changes the target for the "Destroy" button rendered
in `show.html.erb` from a route helper call to a bare model reference,
e.g. from `admin_role_path(@admin_role)` to just `@admin_role`.  This
was missed in #43611.