Fix dbconsole for 3-tier config.
It's possible that applications using a 3-tier config don't have a
"primary" database entry. When booting the db console we should take the
first database, not assume there's a primary.

This also changes the message to be more clear. If no database is
provided and there are no databases for the environment we shouldn't
say that "primary" didn't exist - instead it should say no databases
exist.