Add ReferenceDefinition#add doesn't need a Table argument
When running migrations, any update method called on a Table yielded by
a `change_table` block will ignore the `if_exists` and `if_not_exists`
options, even though migration methods called outside of such a block
will recognize them. This makes it likely that people think they are
adding a conditional table update, when they actually aren't, and
https://github.com/rails/rails/pull/43576 added a feature to raise an
exception when one of the ignored options was passed.

As it happens, `add_reference` uses the same Table class under the hood
which `change_table` also yields, meaning that it's the only method that
can be called outside a change_table block which will also ignore those
options (and also the exceptions that got added in the linked PR were
not properly silenced on migrations of version 7.0 and below).

This commit adds a new ReferenceDefinition method `add` which behaves
identically to `add_to`, but instead of accepting a Table it accepts
a table name and a connection. This will silence the exceptions that
should not have been raised for <7.0 migrations, as well as make the
method properly respond to the if_exist and if_not_exist options.