Automatically inverse_of with custom foreign keys
Background
---

Along the same lines as 518b9a30, this commit allows us to automatically
detect the `inverse_of` for associations with custom foreign_keys by
searching through the reflections on the target class.

For GitHub, this adds `inverse_of` to 161 associations that didn't have
it before (and a total of nearly 300 associations when combined with
`automatic_scope_inversing`). Of those we had 4 associations that relied
on the old behavior and so I added `inverse_of: false` to those for now.

Approach
---

With this commit we first try to find the association by name as before,
but if that doesn't work we can through the reflections of the target
class and look for one that has matching options. To ensure we've really
got a match we have to check just about everything—foreign_key,
primary_key, class, macro, etc.

To ensure we filter out polymorphic reflections earlier (which can blow
up when calling `reflection.foreign_key`), this commit reorders
`valid_inverse_reflection?` a bit, but other callers of that method
shouldn't see any difference in behavior.

Since this is a breaking change (existing code may rely on a missing
`inverse_of` causing fresh copies of a record to be loaded), the
behavior is behind the `automatic_foreign_key_inversing` flag. It is set
to true for new applications via framework defaults.

Alternatives Considered
---

I started off by of chopping off the "_id" part of the foreign key
option to try and guess the inverse name, but that worked for fewer
cases than this does.

Performance considerations:
---

Scanning through reflections to find valid ones could be inefficient on
classes with a large number of associations.

GitHub does have a couple classes with > 100 associations, however the
vast majority have only a few (4-5 on average). Eagerly loading all of
our inverses takes about 160ms longer with this change, which I think is
an acceptable tradeoff for us.