Revert "Merge pull request #43604 from nikonoid/bug/31722"
This reverts commit 2374f0efedda67b316f980b0171320ba0171dd98, reversing
changes made to b37c02e02730133ca85c8479d09b156be46c53d0.

Since this break existing applications, and is necesary only because
applications are not using timestamped migrations and are expecting it
to just work, I decided to not revert this.

Those application should either follow the timestamped migrations format
or disable it.