Short circuit authenticate_by on empty password
Password setter methods ignore empty strings and treat `nil` values
specially.  Thus, empty or `nil` passwords will not be hashed when
passed to a model constructor.  Passing such passwords to
`authenticate_by` results in a large timing difference depending on
whether the model's record is found.

This commit changes `authenticate_by` to return `nil` before querying
the database when any of the given passwords are empty or `nil`.