Refine authenticate_by security guarantee [ci-skip]
We are not in a position to guarantee exactly how long `authenticate_by`
will take.  For example, if the `find_by` query is not backed by an
index, the database will need to perform a full table scan, and the
query time will vary based on where the record is in the heap, if it is
found at all.  Therefore, we should be more specific about what we
guarantee.