Cannot deliver new inbound email via form
1. Install ActionMailbox in your app
2. Visit `/rails/conductor/action_mailbox/inbound_emails` as specified in here https://guides.rubyonrails.org/action_mailbox_basics.html#working-with-action-mailbox-in-development
3. Click on "New inbound email by form"
4. Fill the form and submit

The inbound email is created successfully

A `NoMethodError` is raised:

```
NoMethodError in Rails::Conductor::ActionMailbox::InboundEmailsController#create
undefined method `original_filename' for "":String
```

```
>> mail_params[:attachments]
=> [""]
```

This happens because of the hidden field tag the multiple select field uses under the hood. An easy fix for this problem is to exclude the empty string from the collection.

Add empty string in attachments params in test file

Updated an existing test case to include an empty string in the
attachments collection, mimicking the behavior of the form submission.

Update inbound emails controller test

Moved the empty string for attachments to a different test case where
attachments are not actually sent, just so that we can exercise that
this test fails without this change and passes with it.