Remove requirement that query ends with comments
For the SQL Server adapter
(https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/) the
tests in QueryLogsTest need to be coerced and re-implemented
(https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/972) because the tests
are expecting the SQL to end with the comment but the SQL generated by the
SQL Server adapter ends with prepare statement bindings. This PR just
removes the end of string check in the regular expressions.

These changes do not affect the purpose of the tests and removes the
need coerce & re-implement the same tests in the SQL Server adapter.