Define Thor tasks for test subcommands
This avoids loading the Rakefile, which results in the app booting
twice, once in the development environment to load the Rake tasks, and
then another time in a fork to run the test runner in the test
environment.

Using Thor tasks with no parameters means passing arguments to the task
raises, whereas Rake accepts additional arguments that are not used:

    $ bin/rake test:system SOMETHING=anything

Similarly, with Rake you can use --trace which won't work here.