Fix rollbacks for transactions upon serialization failure or deadlock (#43130)
PR #30922 identified that MySQL cannot handle rollbacks of savepoints
once a serialization failure or deadlock has occurred. It added an
:invalidated state to track this and used it to skip all rollbacks.

With these transactions no longer being rolled back, the underlying
connections began to be thrown away. The connection discard facility
was added in #40541 to address limited issues of certain errors during
rollback.

Unfortunately, the combination of #30922 and #40541 resulted in
throwing away the connection upon every deadlock and serialization
failure--assuredly not what either PR intended.

PostgreSQL allows rollbacks of savepoints and skipping them is only
necessary for MySQL. This commit restores all rollbacks except for
savepoints rollbacks in MySQL.

Due to the interaction of #30922 and #40541, this PR also attempts to
clarify relevant tests and lastly improves tests from #40541 to limit
collateral test failures.