Creating an inbound email in Action Mailbox Conductor raised a NoMethod Error exeception.
mail_params[:attachments] is an array of attachments however the browser is passing the first element as a
blank string whether there are attachments or not. This is causing the NoMethodError as the original_filename
method is being called on the attachment elements.

Resolution is to strip out empty string from attachments array.

See https://github.com/rails/rails/issues/43928