Adjust association inverse_of docs [ci skip]
Before this commit the `inverse_of` note in the `:foreign_key` option
docs was a bit misleading. The text was originally added in [91fd6510][]
to describe adding `inverse_of` to any `belongs_to` associations on a
join model:

> If you are going to modify the association (rather than just read from it), then it is
> a good idea to set the <tt>:inverse_of</tt> option.

That text makes sense for the `:through` option, but it got copied over
to all the `:foreign_key` option docs as well in
https://github.com/rails/rails/pull/31446. Adding `:inverse_of` to an
association with a custom `:foreign_key` is often desirable even if you
are only reading the associations, since it can prevent N+1 queries
(e.g. `post.comments.map(&:post)`).

This commit changes the text to suggest adding `:inverse_of` all the
time when `:foreign_key` is set, rather than only when the association
gets modified.

This commit also points to the `Setting Inverses` section from the
`:through` option docs. `Setting Inverses` directly follows `Association
Join Models` and is perhaps better named `Setting Inverses on
Association Join Models`. It's probably the more relevant part for the
`:through` option docs, but it's easy to miss because of the section
break. This seemed like the easiest fix, but I could also rename the
section or remove the section break entirely.

[91fd6510]: https://github.com/rails/rails/commit/91fd6510563f84ee473bb217bc63ed598abe3f24