Support reading `FormBuilder#id` from `id:` option
When `FormBuilder#id` was introduced in [59ca21c][], it only read the
value passed in as `html: { id: "..." }` and ignored any top-level `id:`
options.

In spite of that fact, since the `<form>` element's is rendered separate
from the builder's block argument, the element still consistently
rendered with an `[id]` attribute. That value is read from the nested
`html: { id: "..." }` option and falling back to any provided `id:
"..."` option.

This commit adds test suite coverage to retain the previous behavior,
along with new tests to support reading `FormBuilder#id` from the `id:`
option.

[59ca21c]: https://github.com/rails/rails/commit/59ca21c0119fff803e75e78d800c0c98aee7fd4c