Stop generating buttons with default type and name
[FormTagHelper#button_tag][] (and transitively, [FormBuilder#button][])
generates a `<button>` with default `[type="submit"]` and
`[name="button"]` attributes.

This commit removes both keys from the hash of default attributes for a
`<button>`.

`[type="submit]`
---

According to the MDN documentation, `<button>` elements will default to
[type="submit"][]:

> `submit`: The button submits the form data to the server. This is the
> default if the attribute is not specified for buttons associated with
> a `<form>`, or if the attribute is an empty or invalid value.

`[name="button"]`
---

Encoding a `[name]` attribute into a `<button>` means that when the
button is clicked, browsers will include the `[name]` (and `[value]`,
when provided) in the submission. In the case of a `[method="get"]`
form, that means the name-value pairing are URL encoded into the
redirect URL. In the case of a `[method="post"]`, that value is URL
encoded (or multi-part form encoded, when an `<input type="file">` is
present) into the body.

It's likely that on the server side, the `button` key is largely ignored
by applications. While the `params[:button]` key is present, the value
is `nil`. Even if the `<button>` element were generated with a `[value]`
attribute, it's unlikely that the pairing would bear any significance in
most applications.

[FormTagHelper#button_tag]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/FormTagHelper.html#method-i-button_tag
[FormBuilder#button]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/FormBuilder.html#method-i-button
[type="submit"]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-type
[name]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-name