Dump the precision for datetime columns following the new defaults
Since Rails 7.0 the datetime column precision is 6 by default.

That means that `t.datetime` calls without setting the `:precision`
option would have its precision set to 6.

The schema dumper was generating a call without precision for columns
with the precision set to `nil` (which has the same effect of 0), making
the schema dump incorrect for the next load since that would be interpreted
as if the precision was 6.

This didn't affect MySQL since #44171, since MySQL precision is 0 by default
not `nil` but it affected PostgreSQL and SQLite3.

Now the dumper will generate the precision as 0 for columns without
precision and omit it when the precision is 6.

Related to #43909.