Improve CSP guide [ci-skip]
"HTTP Security Headers" is more descriptive than "Default Headers", as
these only refer to HTTP Headers.
Add a short introduction explaining "why" a CSP is recommended.
Add headers to the reporting violations and the nonce sections.
Describe what nonce are used for and link to the w3.org CSP documentation.