Add test assertion to ensure Migration::MIGRATOR_SALT cannot be changed.
The MIGRATOR_SALT is a randomly chosen 31-bit signed integer. It is used to multiply the 8-bit unsigned integer produced by the CRC32 hash of the current database being migrated. Together, this results in a 64-bit signed integer which is used to produce the advisory lock id used to orchestrate concurrent migrations. To maintain backwards compatibility across versions, it is important this salt doesn't change. This test provides an extra guarantee that this can't happen, and makes the system more robuste.