Transactions can be restarted
This allows us to simulate an immediately-nested transaction without
ever needing to begin or commit the child: those operations are no-ops,
and we need only run a command if the child rolls back.

Further, we can use a single query to implement that rollback: where the
parent is a savepoint, the standard behaviour is to roll back to
immediately _after_ the savepoint was set. Where the parent is a
top-level transaction, most databases support ROLLBACK AND CHAIN.