Document some Application::Configuration as accessors [ci-skip]
A lot of the undocumented methods in Application::Configuration are
already documented in the configuration guides. Some of these methods
are basically `attr_accessors` so we should document them as such.

This can be done using the `+attr_*+` directives.

As we mark the original method as `:nodoc:` a documentation line above
the original method is required. Otherwise the `+attr_*+` directive is
ingored:

      # :attr_writer: some_attr

      # Required documentation line
      def some_attr=(value) # :nodoc: