Improve serialize API doc [ci-skip]
The "Serialization formats" ASCII table confused the syntax highlighter.
Additionally, the "database storage" column headers were a bit muddled,
and the "NULL" column was difficult to decipher.

This commit relocates that information to the `class_name_or_coder`
parameter description, expressed as a list of accepted values.

This commit also fixes the custom coder code example (which called
`string.rot13` instead of `rot13(string)`), and improves the
documentation formatting in various spots.