Expand HTTP Security Headers in guides [ci-skip]
Adds more details to several HTTP security headers.

The X-Content-Security-Policy header was removed from the list of common
headers as it already has its own section.
The Strict-Transport-Security header was moved from the list of common
headers to its own section.
The Referrer-Policy and X-Permitted-Cross-Domain-Policies headers were
added to the list as they are part of the defaults.

Changed the example for configuring the default headers to overriding
single headers instead of all headers. This could help avoid
unsetting headers by accident.