Add join_existing to transaction so ActiveRecord::Rollback won't be rescued
With join_existing: true then ActiveRecord::Rollback will re-raise
till it reaches a subtransaction or the top transaction.
The reason to stop at subtransaction transactions is that
they are actual nested transactions so are the smallest unit the
database can rollback.

If an ActiveRecord::Rollback is raised inside a nested transaction
then show a deprecation unless:
join_existing: true or a subtransaction was created