Reduce the memory footprint of fixtures accessors
In our large application, the fixtures accessors alone acount for almost 150MB

```
143.79 MB  rails-83fc7402dade/activerecord/lib/active_record/test_fixtures.rb:63
```

While meory usage in test environments is not a huge concern, most
CI environments have memory limits, so it's best not to waste it.

The reason it uses so much memory is that all the accessors are defined
independently with `define_method`, so it creates tons of method entries
and `Proc` instances.

This can be replaced by a proper `method_missing` implementation,
it's very marginally slower, but use a constant amount of memory.