Modernize ActiveSupport::Autoload
We no longer need to define instance variables to avoid uninitialized
instance variable warnings.

Also rather than to keep a hash of `constant -> path` forever, we
can simply keep a list of constants and call `const_get` for each of
them.

And finally we can clear the list we kept, as it's just wasted memory.