Handle net-smtp mail dependency error for Ruby 3.1
In Ruby 3.1, net-stmp is not a default gem anymore. Mail uses it as
dependency and try to loads without declaring the dependency, which
will fail unless the application have net-stmp in the gemfile.

For Rails 7 we could just make those gems dependencies of the framework.
But in Rails 6.1, which we need to support old rubies, installing those
gems can cause trouble. So instead we are intercepting the error and
telling people to add the gem to the Gemfile.