Add Module#eager_load! to trigger the loading of autoloaded constants
It's not uncommon to have gems that are largely autoloaded, which
can be a problem for Rails application in production.

For instance: https://github.com/weppos/whois/pull/644

If the autoloaded constants are not referenced as part of boot
the first request to actually use it will pay the loading cost.
This is what tend to cause "shark fin" looking latency graphs
around deploy.

Additionally to the cost of loading the code, defining constants
cause the Ruby global constant cache to be busted, causing most
inline caches and JITed code to be invalidated.

It also prevent this memory from being shared through the parent
process in forking setups.

`Module#eager_load!` would make it easy to eagerly load some
third party namespaces:

```ruby
config.eager_load_namespaces << Whois::Server::Adapters
```