Add bin/dev and Procfile.dev to new Rails apps
`jsbundling-rails`, `cssbundling-rails`, and `tailwindcss-rails` all introduce a new bin/dev and Procfile.dev file to the app when they are installed. This is a good thing as it allows developers to run one command to execute their Rails server and frontend build tools.

However, this change in flow is not very obvious to newcomers or even Rails developers with even moderate experience. Up until this point, they've been told to simply run `rails s` and they can start developing their app. It's not clear with the introduction of the bundling gems they need to run this separate command to get everything going unless they dig into the gems' README files. They are then left wondering why none of their Javascript or CSS changes are applying because its not obvious they have to run a separate process.

This change introduces `bin/dev` and `Procfile.dev` to Rails proper as first class citizens added on the defualt app install - regardless of `--javascript` or `--css` flags. A `--skip-bin-dev` flag was added to not generate those files.

`bin/dev` is identical to the one found in the bundling gems. Their installers will see the identical file and skip it. `Procfile.dev` includes only the "web" process which is also taken from the bundling gems.

So why even have `Procfile.dev` if it only has one service in it? If the default workflow of Rails apps going forward uses `bin/dev` to start/stop the dev stack, this can result in a more smoother transition from importmaps/raw css into the bundling gems as developers would already be using `bin/dev`. The gems would append to `Procfile.dev` and no other change would be needed on the developer's part. Also, this allows for other gems such as `sidekiq`, `resque`, `shackapacker`, `goodjob`, etc (which also need their own separate process while working locally) to slide into the Rails flow easily as one would append a new service to `Procfile.dev`.